+++ client/package.json
@@ -13,7 +13,8 @@
     "axios": "^1.3.6",
     "firebase": "^9.22.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
